refactor: update generateProfileUrl types
