+++ lib/main.ts
@@ -17,7 +17,6 @@ export {
   frameworkSettings,
   splitString,
   generateProfileUrl,
-  ProfilePage,
 } from "./utils";
 
 export {
